requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
bokeh==3.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
configobj==5.0.8
configparser==7.0.0
contourpy==1.2.1
dataclasses-json==0.6.6
debugpy==1.8.1
decorator==5.1.1
distro==1.9.0
duckduckgo_search==6.1.4
et-xmlfile==1.1.0
etelemetry==0.3.1
executing==2.0.1
filelock==3.14.0
frontend==0.0.3
frozenlist==1.4.1
fsspec==2024.5.0
greenlet==3.0.3
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.2
idna==3.7
intel-openmp==2021.4.0
ipykernel==6.29.4
ipython==8.24.0
ipywidgets==8.1.3
isodate==0.6.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_widgets==3.0.11
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.2
langchain-experimental==0.0.59
langchain-huggingface==0.0.1
langchain-openai==0.1.8
langchain-text-splitters==0.2.0
langgraph==0.0.59
langsmith==0.1.65
linkify-it-py==2.0.3
looseversion==1.3.0
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.1
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
neo4j==5.20.0
nest-asyncio==1.6.0
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
numpy==1.26.4
openai==1.30.5
openpyxl==3.1.3
orjson==3.10.3
packaging==23.2
pandas==2.2.2
panel==1.4.3
param==2.1.0
parso==0.8.4
pathlib==1.0.1
pillow==10.3.0
platformdirs==4.2.2
prompt_toolkit==3.0.45
protobuf==5.27.0
prov==2.0.0
psutil==5.9.8
pure-eval==0.2.2
pydantic==2.7.2
pydantic_core==2.18.3
pydot==2.0.0
Pygments==2.18.0
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
pyparsing==3.1.2
pyreqwest_impersonate==0.4.7
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyviz_comms==3.0.2
pywin32==306
pyxnat==1.6.2
PyYAML==6.0.1
pyzmq==26.0.3
rdflib==7.0.0
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
sentence-transformers==3.0.0
sentencepiece==0.2.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
stack-data==0.6.3
starlette==0.37.2
sympy==1.12.1
tbb==2021.12.0
tenacity==8.3.0
tesseract==0.1.3
text-generation==0.7.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.3.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
traits==6.3.2
transformers==4.41.1
typing-inspect==0.9.0
typing_extensions==4.12.0
tzdata==2024.1
uc-micro-py==1.0.3
urllib3==2.2.1
uvicorn==0.30.0
wcwidth==0.2.13
webencodings==0.5.1
widgetsnbextension==4.0.11
wikipedia==1.4.0
xyzservices==2024.4.0
yarl==1.9.4
yfiles_jupyter_graphs==1.6.2
